Move mixt.pyxl.codec to mixt.codec and remove pyxl directory
