+++ .pylintrc
@@ -9,7 +9,7 @@
 
 # Add files or directories to the blacklist. They should be base names, not
 # paths.
-ignore=pyxl
+ignore=codec
 
 # Pickle collected data for later comparisons.
 persistent=yes
