+++ setup.cfg
@@ -80,7 +80,7 @@ python_version = 3.6
 strict_optional = true
 warn_incomplete_stub = true
 
-[mypy-src.mixt.pyxl.*]
+[mypy-src.mixt.codec.*]
 # ignore imported pyxl code
 ignore_errors = True
 
@@ -131,7 +131,7 @@ exclude =
     *.egg-info
     dist/
     build/
-    src/mixt/pyxl/
+    src/mixt/codec/
     tests/pyxl/original/
 per-file-ignores =
     # ignore mypy missing annotations in tests
