+++ src/mixt/codec/README.md
@@ -1,0 +1,3 @@
+This is from `pyxl.codec`, updated for our needs.
+
+The original code is here: https://github.com/pyxl3/pyxl3 (the `fixtests` branch at the time of the copy).
