--- src/mixt/pyxl/codec/__init__.py
@@ -1,1 +1,0 @@
-#!/usr/bin/env python
