--- src/mixt/pyxl/example.py
@@ -1,5 +1,0 @@
-# coding: mixt
-
-from mixt import html
-
-print(<html><body>Hello World!</body></html>)
